Move file importing into a real Action

+++ src/PortfolioActions.js
@@ -3,32 +3,18 @@
 import * as React from 'react';
 import { Button, Col, Input, Label } from 'reactstrap';
 import type { Dispatch } from './types';
-import { addTransactions, deletePortfolio, downloadPortfolio, fetchAllQuotes } from './actions';
+import { deletePortfolio, downloadPortfolio, importTransactionsFile } from './actions';
 import { connect } from 'react-redux';
-import csvParse from 'csv-parse/lib/es5/sync';
-import { transformGfToStocks } from './transformers';
 
 type Props = { dispatch: Dispatch };
 
-type State = {
-  isReadingFile: boolean,
-};
-
-class PortfolioActions extends React.Component<Props, State> {
-  fileReader: ?FileReader;
-
-  constructor(props: Props) {
-    super(props);
-    this.state = { isReadingFile: false };
-  }
-
-  componentWillUnmount() {
-    // Ensure any in-flight file reading is killed if this component is about to be unmounted.
-    if (this.fileReader != null) this.fileReader.onload = function() {};
-  }
-
+class PortfolioActions extends React.Component<Props> {
   handleDeletePortfolio = () => {
-    this.props.dispatch(deletePortfolio());
+    if (
+      window.confirm('Are you sure you want to delete the entire portfolio? This is irreversible.')
+    ) {
+      this.props.dispatch(deletePortfolio());
+    }
   };
 
   handleDownloadPortfolio = () => {
@@ -39,29 +25,18 @@ class PortfolioActions extends React.Component<Props, State> {
     const currentTarget = event.currentTarget;
     const files = currentTarget.files;
     if (files == null || files.length === 0) return;
+    this.props.dispatch(importTransactionsFile(files[0]));
 
-    const file = files[0];
-    this.setState({ isReadingFile: true }, () => {
-      const fileReader = (this.fileReader = new FileReader());
-      fileReader.onload = () => {
-        const parsedCsv = csvParse(fileReader.result, { columns: true });
-        this.props.dispatch(addTransactions(transformGfToStocks(parsedCsv)));
-        this.props.dispatch(fetchAllQuotes());
-        this.setState({ isReadingFile: false });
-
-        // Reset the input so the same file can be uploaded multiple times in a row (without
-        // resetting the `onchange` would not fire). Why upload multiple times? Testing testing
-        // testing. ABT: Always Be Testing.
-        currentTarget.value = '';
-      };
-      fileReader.readAsText(file);
-    });
+    // Reset the input so the same file can be uploaded multiple times in a row (without
+    // resetting the `onchange` would not fire). Why upload multiple times? Testing testing
+    // testing. ABT: Always Be Testing.
+    currentTarget.value = '';
   };
 
   render() {
     return (
       <Col className="text-right">
-        <Button color="link" disabled={this.state.isReadingFile} size="sm" type="button">
+        <Button color="link" size="sm" type="button">
           <Label className="label-button">
             <Input accept="text/csv" hidden onChange={this.handleImportTransactions} type="file" />
             Import transactions
